requirements.txt
blessings==1.6
bpython==0.14.2
curtsies==0.1.19
Cython==0.23.3
Django==1.8.5
django-bower==5.0.4
django-filter==0.11.0
django-pipeline==1.5.4
django-websocket-redis==0.4.4
djangorestframework==3.2.4
futures==3.0.3
gevent==1.1b6.dev0
greenlet==0.4.9
Markdown==2.6.2
Pygments==2.0.2
redis==2.10.3
requests==2.8.0
six==1.10.0
wheel==0.24.0
